requirements.txt
streamlit
google-generativeai
beautifulsoup4
lxml
scikit-learn
fpdf
requests
python-dotenv
tqdm
